数値に対応

+++ SpaceWars2/scenes/Game.hpp
@@ -12,8 +12,8 @@ private:
 	LineString RHPGraph;
 
 	static void rightAlign(String _font, String _text, int _x, int _y, Color _color = Color(L"#fff"));
-	static Quad gauge(bool _isLeft, Vec2 _topLeft, int _width);
-	static void drawGauge(bool _isLeft, Vec2 _topLeft, int _width, String _fillColor, String _frameColor);
+	static Quad gauge(bool _isLeft, Vec2 _topLeft, double _width);
+	static void drawGauge(bool _isLeft, Vec2 _topLeft, double _width, String _fillColor, String _frameColor);
 
 public:
 	void init() override;
